Avoid resetting pending frame
The function update_texture() calls update_texture_internal() and falls
back to set_pending_frame() if it fails.

When the frame passed is the pending frame, call only the _internal()
version instead.

This will prevent issues with frame reference counts by ensuring the
source and destination frames are never the same.

Refs 6298ef095ffa9a2cdd2b4d245e71280743b5a59e
Refs #6357 <https://github.com/Genymobile/scrcpy/issues/6357>